HW4: comments author + gates (edit/delete for author/moderator)

+++ app/Http/Controllers/CommentController.php
@@ -5,17 +5,14 @@ namespace App\Http\Controllers;
 use App\Models\Article;
 use App\Models\Comment;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 
 class CommentController extends Controller
 {
-    public function __construct()
-    {
-        $this->authorizeResource(Comment::class, 'comment');
-    }
-
     public function index(Article $article)
     {
         $comments = $article->comments()
+            ->with('author')
             ->latest()
             ->paginate(10);
 
@@ -27,6 +24,8 @@ class CommentController extends Controller
 
     public function create(Article $article)
     {
+        Gate::authorize('comment.create');
+
         return view('comments.create', [
             'article' => $article,
             'comment' => new Comment(),
@@ -35,13 +34,22 @@ class CommentController extends Controller
 
     public function store(Request $request, Article $article)
     {
+        Gate::authorize('comment.create');
+
         $data = $request->validate([
-            'author_name'  => ['required', 'string', 'min:2', 'max:100'],
-            'author_email' => ['nullable', 'email', 'max:150'],
-            'body'         => ['required', 'string', 'min:3', 'max:2000'],
+            'body' => ['required', 'string', 'min:3', 'max:2000'],
         ]);
 
-        $article->comments()->create($data);
+        $user = $request->user();
+
+        $article->comments()->create([
+            'user_id' => $user->id,
+            // Дублируем имя/почту для удобного вывода (и чтобы комментарий не "сломался",
+            // если у пользователя поменяются данные)
+            'author_name' => $user->name,
+            'author_email' => $user->email,
+            'body' => $data['body'],
+        ]);
 
         return redirect()
             ->route('articles.show', $article)
@@ -50,6 +58,8 @@ class CommentController extends Controller
 
     public function edit(Comment $comment)
     {
+        Gate::authorize('comment.update', $comment);
+
         $comment->load('article');
 
         return view('comments.edit', [
@@ -60,13 +70,15 @@ class CommentController extends Controller
 
     public function update(Request $request, Comment $comment)
     {
+        Gate::authorize('comment.update', $comment);
+
         $data = $request->validate([
-            'author_name'  => ['required', 'string', 'min:2', 'max:100'],
-            'author_email' => ['nullable', 'email', 'max:150'],
-            'body'         => ['required', 'string', 'min:3', 'max:2000'],
+            'body' => ['required', 'string', 'min:3', 'max:2000'],
         ]);
 
-        $comment->update($data);
+        $comment->update([
+            'body' => $data['body'],
+        ]);
 
         return redirect()
             ->route('articles.show', $comment->article)
@@ -75,6 +87,8 @@ class CommentController extends Controller
 
     public function destroy(Comment $comment)
     {
+        Gate::authorize('comment.delete', $comment);
+
         $article = $comment->article;
 
         $comment->delete();
